Small moving in transformers
